Encode google API key with Base64 (#9807)
* encode the google api-key before passing into JSON

* updated the test cases

* resolved a few issues

* resolved a few issues

* resolved a few issues

* made changes as suggested

* resolved a few issues

* resolved a few issues

* resolved a few issues

* resolved a minor define error

* resolved some minor Flake8 errors

* renamed regex.py into string.py

* renamed test_string.py to test_string_regex.py

* fixed a few minor issues

* renamed test file to test_string_properties.py

* updated gmap_plot_canvas.ts
