+++ bokeh/plotting/gmap.py
@@ -232,7 +232,6 @@ def gmap(google_api_key, map_options, **kwargs):
        GMap
 
     '''
-
     return GMap(api_key=google_api_key, map_options=map_options, **kwargs)
 
 #-----------------------------------------------------------------------------
