+++ bokehjs/src/lib/models/plots/gmap_plot_canvas.ts
@@ -52,7 +52,8 @@ export class GMapPlotView extends PlotView {
 
     if (typeof google === "undefined" || google.maps == null) {
       if (typeof window._bokeh_gmaps_callback === "undefined") {
-        load_google_api(this.model.api_key)
+        const decoded_api_key = atob(this.model.api_key)
+        load_google_api(decoded_api_key)
       }
       gmaps_ready.connect(() => this.request_render())
     }
