+++ tests/unit/bokeh/core/property/test_string_properties.py
@@ -19,7 +19,7 @@ from _util_property import _TestHasProps, _TestModel
 from bokeh._testing.util.api import verify_all
 
 # Module under test
-import bokeh.core.property.regex as bcpr # isort:skip
+import bokeh.core.property.string as bcpr # isort:skip
 
 #-----------------------------------------------------------------------------
 # Setup
@@ -27,6 +27,7 @@ import bokeh.core.property.regex as bcpr # isort:skip
 
 ALL = (
     'Regex',
+    'Base64String',
 )
 
 #-----------------------------------------------------------------------------
